latest updates to v4 batches 03/24/2023

+++ alpha-fold/re-download/dat-to-out.py
@@ -6,7 +6,7 @@
 # to cater for cases where line.split() does not contain 12 items.
 # Added a try block for cases where .dat files exist instead of .dssp
 
-import sys
+import sys, os
 import numpy
 
 name=sys.argv[1]
@@ -15,26 +15,29 @@ pdb_name=name+".pdb"
 ss_name=name+".dssp"
 out_name=name+".out"
 
-try:
-    file_ss=open(ss_name,"r")
-except:
-    ss_name=name+".dat"
-    file_ss=open(ss_name,"r")
+if os.path.isfile(out_name):
+    print(f"{out_name} already exists")
+else:
+  try:
+      file_ss=open(ss_name,"r")
+  except:
+      ss_name=name+".dat"
+      file_ss=open(ss_name,"r")
 
-file_out=open(out_name,"w+")
+  file_out=open(out_name,"w+")
 
-# skip the first 28 line for the secondary structure file
-for i in range(28):
-  file_ss.readline()
+  # skip the first 28 line for the secondary structure file
+  for i in range(28):
+    file_ss.readline()
 
-with open(pdb_name) as file_pdb:
-  for line in file_pdb:
-    if line.split()[0]=='ATOM':
-      if (line.split()[2]=='CA'):
-        ss=file_ss.readline()[16]
-        # print(file_ss.readline())
-        # print(ss)
-        if (ss==' '):
-          ss='~'
-        
-        print(line.split()[3],line.split()[-2],ss,file=file_out)
+  with open(pdb_name) as file_pdb:
+    for line in file_pdb:
+      if line.split()[0]=='ATOM':
+        if (line.split()[2]=='CA'):
+          ss=file_ss.readline()[16]
+          # print(file_ss.readline())
+          # print(ss)
+          if (ss==' '):
+            ss='~'
+          
+          print(line.split()[3],line.split()[-2],ss,file=file_out)
